Move substitution logic out of fetch_cvd.cc so it can be reused for `start`
