+++ base/cvd/cuttlefish/host/commands/cvd/fetch/BUILD.bazel
@@ -16,6 +16,22 @@ cc_proto_library(
 )
 
 cc_library(
+    name = "substitute",
+    srcs = ["substitute.cc"],
+    hdrs = ["substitute.h"],
+    copts = COPTS,
+    strip_include_prefix = "//cuttlefish",
+    deps = [
+        ":host_pkg_migration_cc_proto",
+        "//cuttlefish/common/libs/utils",
+        "//cuttlefish/common/libs/utils:result",
+        "//libbase",
+        "@fmt",
+        "@protobuf",
+    ],
+)
+
+cc_library(
     name = "fetch",
     srcs = [
         "fetch_cvd.cc",
@@ -30,7 +46,7 @@ cc_library(
     copts = COPTS + [ "-Werror=sign-compare" ],
     strip_include_prefix = "//cuttlefish",
     deps = [
-        ":host_pkg_migration_cc_proto",
+        ":substitute",
         "//cuttlefish/common/libs/utils",
         "//cuttlefish/common/libs/utils:environment",
         "//cuttlefish/common/libs/utils:result",
@@ -45,7 +61,6 @@ cc_library(
         "@fmt",
         "@gflags",
         "@jsoncpp",
-        "@protobuf",
     ],
 )
