+++ base/cvd/cuttlefish/host/commands/cvd/fetch/substitute.h
@@ -1,0 +1,32 @@
+//
+// Copyright (C) 2019 The Android Open Source Project
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include <string>
+#include <vector>
+
+#include "common/libs/utils/result.h"
+
+namespace cuttlefish {
+
+Result<void> SubstituteWithFlag(
+    const std::string& target_dir,
+    const std::vector<std::string>& debian_substitutions);
+
+Result<void> SubstituteWithMarker(const std::string& target_dir,
+                                  const std::string& marker_file);
+
+}  // namespace cuttlefish
